Moved template validation from rules to rulefmt
